Post game-jam polishes
- Tweaked anchors of UI boxes so they work in web-browser.
- Filled in boundaries of map with sand tiles. Forgot to do that before.
- Stopped Skrypers spawning below the map.
- Put coins in a few more places.

Unfortunately wasn't able to submit in time. I'd published the itch.io page
on my account but when I clicked Submit on the Kenney Jam page, it said
there were no projects on my Dashboard that could be submitted. Tried
lots of things but ran out of time in the end. Bit of a bummer. The game
as of this commit is uploaded to itch.io though, as a In-browser game.